refactor index
