+++ src/utils/transformations.ts
@@ -1,4 +1,7 @@
-import { calculateStdDev, getTopValues } from "./utils/functions.js";
+import { calculateStdDev, getTopValues, TopValueItem } from "./functions.js";
+import { QueryResultValue } from "../types/query.js";
+import { z } from "zod";
+import { QueryToolSchema } from "../types/schema.js";
 
 /**
  * Types for the summary statistics
@@ -22,7 +25,7 @@ interface CountStats {
 
 interface BreakdownStat {
   uniqueCount: number;
-  topValues?: Array<{value: any, count: number}>;
+  topValues?: Array<TopValueItem>;
 }
 
 interface BreakdownStats {
@@ -40,7 +43,7 @@ interface ResultSummary {
  * Calculate summary statistics for query results to provide useful insights
  * without overwhelming the context window
  */
-export function summarizeResults(results: any[], params: any): ResultSummary {
+export function summarizeResults(results: QueryResultValue[], params: z.infer<typeof QueryToolSchema>): ResultSummary {
   if (!results || results.length === 0) {
     return { count: 0 };
   }
@@ -52,17 +55,25 @@ export function summarizeResults(results: any[], params: any): ResultSummary {
   // If we have calculation columns, add some statistics about them
   if (params.calculations) {
     const numericColumns = params.calculations
-      .filter((calc: any) => 
+      .filter(calc => 
         calc.op !== "COUNT" && 
         calc.op !== "CONCURRENCY" && 
         calc.op !== "HEATMAP" &&
         calc.column
       )
-      .map((calc: any) => `${calc.op}(${calc.column})`);
+      .map(calc => `${calc.op}(${calc.column})`);
     
     numericColumns.forEach((colName: string) => {
       if (results[0] && colName in results[0]) {
-        const values = results.map(r => r[colName]).filter(v => v !== null && v !== undefined);
+        // Filter to ensure we only have numeric values
+        const values = results
+          .map(r => r[colName])
+          .filter((v): v is number => 
+            v !== null && 
+            v !== undefined && 
+            typeof v === 'number'
+          );
+          
         if (values.length > 0) {
           const min = Math.min(...values);
           const max = Math.max(...values);
@@ -71,10 +82,27 @@ export function summarizeResults(results: any[], params: any): ResultSummary {
           
           // Calculate median (P50 approximation)
           const sortedValues = [...values].sort((a, b) => a - b);
-          const medianIndex = Math.floor(sortedValues.length / 2);
-          const median = sortedValues.length % 2 === 0
-            ? (sortedValues[medianIndex - 1] + sortedValues[medianIndex]) / 2
-            : sortedValues[medianIndex];
+          
+          // Default to average if we can't calculate median properly
+          let median = avg;
+          
+          // We know values is not empty at this point because we checked values.length > 0 earlier
+          if (sortedValues.length === 1) {
+            median = sortedValues[0]!;
+          } else if (sortedValues.length > 1) {
+            const medianIndex = Math.floor(sortedValues.length / 2);
+            
+            if (sortedValues.length % 2 === 0) {
+              // Even number of elements - average the middle two
+              // We can use non-null assertion (!) because we know these indices exist
+              // when sortedValues.length > 1 and we're in the even case
+              median = (sortedValues[medianIndex - 1]! + sortedValues[medianIndex]!) / 2;
+            } else {
+              // Odd number of elements - take the middle one
+              // We can use non-null assertion (!) because we know this index exists
+              median = sortedValues[medianIndex]!;
+            }
+          }
           
           // Now properly typed
           summary[colName] = { 
@@ -91,9 +119,17 @@ export function summarizeResults(results: any[], params: any): ResultSummary {
     });
     
     // Special handling for COUNT operations
-    const hasCount = params.calculations.some((calc: any) => calc.op === "COUNT");
-    if (hasCount && results.length > 0 && 'COUNT' in results[0]) {
-      const countValues = results.map(r => r.COUNT).filter(v => v !== null && v !== undefined);
+    const hasCount = params.calculations.some(calc => calc.op === "COUNT");
+    if (hasCount && results.length > 0 && 'COUNT' in results[0]!) {
+      // Filter to ensure we only have numeric values
+      const countValues = results
+        .map(r => r.COUNT)
+        .filter((v): v is number => 
+          v !== null && 
+          v !== undefined && 
+          typeof v === 'number'
+        );
+        
       if (countValues.length > 0) {
         const totalCount = countValues.reduce((a, b) => a + b, 0);
         const maxCount = Math.max(...countValues);
